+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 
 ## 20260315-0
 
+Simple bugfix release.
+
 * Fix random play popover theming.
 * [#73](https://github.com/RobertWesner/YouTube-Play-All/issues/73) Fix issues with random play logic introduced by refactors in version 20251115-0.
+* Fix mobile button theming.
 
 
 ## Infrastructure 2026-03-02